enp/requirements_enrichment.txt
psycopg2-binary==2.9.9
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
torch>=2.0.0
